move flags-static logic to flags class
