+++ app/src/main/java/com/trianguloy/urlchecker/modules/list/OpenModule.java
@@ -17,6 +17,7 @@ import com.trianguloy.urlchecker.modules.AModuleConfig;
 import com.trianguloy.urlchecker.modules.AModuleData;
 import com.trianguloy.urlchecker.modules.AModuleDialog;
 import com.trianguloy.urlchecker.modules.companions.CTabs;
+import com.trianguloy.urlchecker.modules.companions.Flags;
 import com.trianguloy.urlchecker.modules.companions.LastOpened;
 import com.trianguloy.urlchecker.url.UrlData;
 import com.trianguloy.urlchecker.utilities.AndroidUtils;
@@ -251,7 +252,7 @@ class OpenDialog extends AModuleDialog {
         }
 
         // Get flags from global data (probably set by flags module, if active)
-        Integer flags = FlagsDialog.getGlobalFlagsNullable(this);
+        Integer flags = Flags.getGlobalFlagsNullable(this);
         if (flags != null) {
             intent.setFlags(flags);
         }
